+++ Asserv2015.X/lib_asserv/private/motion.c
@@ -137,7 +137,7 @@ void check_blocked(Speed speed,Speed order){
 
 
 // renvoie les commandes des roues gauche et droite (appelé par l'interruption toutes les 5 ms)
-void motion_step(float tics_g, float tics_d, float *commande_g, float *commande_d){
+void motion_step(int tics_g, int tics_d, float *commande_g, float *commande_d){
     if (!motion_initialized){*commande_g = 0; *commande_d = 0;}
     else {
         // maj de l'odométrie
